+++ db/src/main/java/com/jenschen/dao/BaseDao.java
@@ -37,7 +37,7 @@ public interface BaseDao<T> {
      *
      * @param queryWrapper 实体对象封装操作类（可以为 null）
      */
-    Integer selectCount(@Param(Constants.WRAPPER) Wrapper<T> queryWrapper);
+    Long selectCount(@Param(Constants.WRAPPER) Wrapper<T> queryWrapper);
 
     /**
      * 根据 entity 条件，查询全部记录
