+++ manager-service/src/main/java/com/jenschen/response/PageResp.java
@@ -6,11 +6,11 @@ import java.util.List;
 
 @Data
 public class PageResp<T> {
-    private int total;
+    private long total;
     private List<T> data;
 
 
-    public static <T> PageResp<T> build(int total, List<T> data){
+    public static <T> PageResp<T> build(long total, List<T> data){
         PageResp<T> pageResp = new PageResp<>();
         pageResp.setData(data);
         pageResp.setTotal(total);
